Try to simplify farao-dichotomy lib
