--- dichotomy-api/src/main/java/com/farao_community/farao/dichotomy/api/DichotomyException.java
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2020, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.farao_community.farao.dichotomy.api;
-
-/**
- * @author Sebastien Murgey {@literal <sebastien.murgey at rte-france.com>}
- */
-public class DichotomyException extends RuntimeException {
-    public DichotomyException(String message) {
-        super(message);
-    }
-}
