+++ farao-dichotomy-shift/src/test/resources/com/farao_community/farao/dichotomy/shift/20210901_2230_213_GSK_CO_CSE1.xml
