--- network-dichotomy/src/test/java/com/farao_community/farao/dichotomy/network/RaoResultMock.java
@@ -1,145 +1,0 @@
-package com.farao_community.farao.dichotomy.network;
-
-import com.farao_community.farao.commons.Unit;
-import com.farao_community.farao.data.crac_api.State;
-import com.farao_community.farao.data.crac_api.cnec.FlowCnec;
-import com.farao_community.farao.data.crac_api.network_action.NetworkAction;
-import com.farao_community.farao.data.crac_api.range_action.PstRangeAction;
-import com.farao_community.farao.data.crac_api.range_action.RangeAction;
-import com.farao_community.farao.data.rao_result_api.ComputationStatus;
-import com.farao_community.farao.data.rao_result_api.OptimizationState;
-import com.farao_community.farao.data.rao_result_api.RaoResult;
-import org.apache.commons.lang3.NotImplementedException;
-
-import java.util.Map;
-import java.util.Set;
-
-/**
- * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
- */
-public class RaoResultMock implements RaoResult {
-    private final double cost;
-
-    public RaoResultMock(boolean secure) {
-        this.cost = secure ? -10 : 10;
-    }
-
-    @Override
-    public ComputationStatus getComputationStatus() {
-        return null;
-    }
-
-    @Override
-    public double getFlow(OptimizationState optimizationState, FlowCnec flowCnec, Unit unit) {
-        return 0;
-    }
-
-    @Override
-    public double getMargin(OptimizationState optimizationState, FlowCnec flowCnec, Unit unit) {
-        return 0;
-    }
-
-    @Override
-    public double getRelativeMargin(OptimizationState optimizationState, FlowCnec flowCnec, Unit unit) {
-        return 0;
-    }
-
-    @Override
-    public double getCommercialFlow(OptimizationState optimizationState, FlowCnec flowCnec, Unit unit) {
-        return 0;
-    }
-
-    @Override
-    public double getLoopFlow(OptimizationState optimizationState, FlowCnec flowCnec, Unit unit) {
-        return 0;
-    }
-
-    @Override
-    public double getPtdfZonalSum(OptimizationState optimizationState, FlowCnec flowCnec) {
-        return 0;
-    }
-
-    @Override
-    public double getCost(OptimizationState optimizationState) {
-        return 0;
-    }
-
-    @Override
-    public double getFunctionalCost(OptimizationState optimizationState) {
-        if (optimizationState == OptimizationState.AFTER_CRA) {
-            return cost;
-        } else {
-            throw new NotImplementedException("Only after CRA is handled");
-        }
-    }
-
-    @Override
-    public double getVirtualCost(OptimizationState optimizationState) {
-        return 0;
-    }
-
-    @Override
-    public Set<String> getVirtualCostNames() {
-        return null;
-    }
-
-    @Override
-    public double getVirtualCost(OptimizationState optimizationState, String s) {
-        return 0;
-    }
-
-    @Override
-    public boolean wasActivatedBeforeState(State state, NetworkAction networkAction) {
-        return false;
-    }
-
-    @Override
-    public boolean isActivatedDuringState(State state, NetworkAction networkAction) {
-        return false;
-    }
-
-    @Override
-    public Set<NetworkAction> getActivatedNetworkActionsDuringState(State state) {
-        return null;
-    }
-
-    @Override
-    public boolean isActivatedDuringState(State state, RangeAction rangeAction) {
-        return false;
-    }
-
-    @Override
-    public int getPreOptimizationTapOnState(State state, PstRangeAction pstRangeAction) {
-        return 0;
-    }
-
-    @Override
-    public int getOptimizedTapOnState(State state, PstRangeAction pstRangeAction) {
-        return 0;
-    }
-
-    @Override
-    public double getPreOptimizationSetPointOnState(State state, RangeAction rangeAction) {
-        return 0;
-    }
-
-    @Override
-    public double getOptimizedSetPointOnState(State state, RangeAction rangeAction) {
-        return 0;
-    }
-
-    @Override
-    public Set<RangeAction> getActivatedRangeActionsDuringState(State state) {
-        return null;
-    }
-
-    @Override
-    public Map<PstRangeAction, Integer> getOptimizedTapsOnState(State state) {
-        return null;
-    }
-
-    @Override
-    public Map<RangeAction, Double> getOptimizedSetPointsOnState(State state) {
-        return null;
-    }
-}
