+++ pom.xml
@@ -16,11 +16,6 @@
     <description>Parent pom for dichotomy related FARAO processes</description>
     <packaging>pom</packaging>
 
-    <modules>
-        <module>dichotomy-api</module>
-        <module>network-dichotomy</module>
-    </modules>
-
     <url>http://www.farao-community.com</url>
 
     <licenses>
@@ -45,11 +40,16 @@
         </developer>
     </developers>
 
+    <modules>
+        <module>farao-dichotomy-api</module>
+        <module>farao-dichotomy-shift</module>
+    </modules>
+
     <properties>
         <logback.classic.version>1.2.3</logback.classic.version>
         <mockito.core.version>3.5.10</mockito.core.version>
+        <farao.bom.version>1.1.0</farao.bom.version>
     </properties>
-
 
     <dependencyManagement>
         <dependencies>
@@ -62,6 +62,13 @@
                 <groupId>org.mockito</groupId>
                 <artifactId>mockito-core</artifactId>
                 <version>${mockito.core.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.farao-community.farao</groupId>
+                <artifactId>farao-bom</artifactId>
+                <version>${farao.bom.version}</version>
+                <scope>import</scope>
+                <type>pom</type>
             </dependency>
         </dependencies>
     </dependencyManagement>
